requirements.txt
numpy
pandas
matplotlib
seaborn
requests
python-dotenv
ipykernel
xgboost
scikit-learn
